# libs/client/rest/src/nutri-openapi.yaml
openapi: 3.0.0
paths:
...
  /api/recipes:
    post:
      operationId: RecipesController_create
      summary: Create a new recipe
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecipeDto'
      responses:
        '201':
          description: Recipe successfully created
        '400':
          description: Bad request
      tags:
        - Recipe Management
    get:
      operationId: RecipesController_findAll
      summary: Find all recipes
      parameters:
        - name: search
          required: false
          in: query
          description: Search term for filtering recipes
          schema:
            type: string
      responses:
        '200':
          description: List of recipes returned
      tags:
        - Recipe Management
  /api/recipes/{id}:
    get:
      operationId: RecipesController_findOne
      summary: Find a recipe by ID
      parameters:
        - name: id
          required: true
          in: path
          description: Recipe ID
          schema:
            type: string
      responses:
        '200':
          description: Recipe found
        '404':
          description: Recipe not found
      tags:
        - Recipe Management
    delete:
      operationId: RecipesController_remove
      summary: Delete a recipe
      parameters:
        - name: id
          required: true
          in: path
          description: Recipe ID
          schema:
            type: string
      responses:
        '200':
          description: Recipe successfully deleted
        '404':
          description: Recipe not found
      tags:
        - Recipe Management
  /api/recipes/{id}/full:
    get:
      operationId: RecipesController_getFullRecipe
      summary: Get full recipe details including ingredients and instructions
      parameters:
        - name: id
          required: true
          in: path
          description: Recipe ID
          schema:
            type: string
      responses:
        '200':
          description: Complete recipe details returned
        '404':
          description: Recipe not found
      tags:
        - Recipe Management
  /api/recipes/{id}/cooking-view:
    get:
      operationId: RecipesController_getCookingView
      summary: Get recipe in cooking view format
      parameters:
        - name: id
          required: true
          in: path
          description: Recipe ID
          schema:
            type: string
      responses:
        '200':
          description: Cooking view of recipe returned
        '404':
          description: Recipe not found
      tags:
        - Recipe Management
  /api/recipes/{id}/parts:
    post:
      operationId: RecipesController_createPart
      summary: Create a new part for a recipe
      parameters:
        - name: id
          required: true
          in: path
          description: Recipe ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Recipe part created successfully
        '404':
          description: Recipe not found
      tags:
        - Recipe Management
  /api/schemas/recipe/{operation}:
    get:
      operationId: RecipeSchemasController_getRecipeSchema
      summary: Get recipe JSON schema for specific operation
      parameters:
        - name: operation
          required: true
          in: path
          description: Operation type (create, update, etc.)
          schema:
            type: string
      responses:
        '200':
          description: Schema returned successfully
      tags:
        - Recipe Schemas
...
info:
  title: Nutri API
  description: Internal REST API documentation
  version: 0.0.1
  contact: {}
...
servers: []
components:
  schemas:
...
    CreateRecipeDto:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
        sourceId:
          type: string
        sourceUrl:
          type: string
        cookingTime:
          type: string
        prepTime:
          type: string
        servingsText:
          type: string
        servingsMin:
          type: number
          minimum: 1
        servingsMax:
          type: number
          minimum: 1
        tagIds:
          type: array
          items:
            type: string
      required:
        - title
...